query all projects at once
This commit changes the graphql query to query all
projects at once. This query is a lot faster than
the per-project queries, but is limited to 100
projects, at most 100 items per proejct and at
most 48 fields per item.

Additional nodes have to be retrieved with
additional queries. These additional queries
are not part of this commit.

The item index entry has been removed